requirements.txt
redis>=7.1.0
fastapi==0.128.0
uvicorn==0.40.0
httpx==0.28.1
pydantic-settings==2.11.0
pydantic==2.12.5
tenacity==9.1.2
colorlog==6.10.1
dacite==1.9.2
dataclasses-json==0.6.7
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.2.1